move fn back over to metricQuery
